server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
python-jose[cryptography]==3.3.0  # For JWT tokens
passlib[bcrypt]==1.7.4  # For password hashing
